updated solo serve
