+++ solo_server/commands/serve.py
@@ -1,17 +1,18 @@
 import typer
 import os
 import json
-import click
 import subprocess
-import time
-from enum import Enum
-from pathlib import Path
 
+from enum import Enum
+from typing import Optional
 from solo_server.config import CONFIG_PATH
 from solo_server.config.config_loader import get_server_config
 from solo_server.utils.hardware import detect_hardware
-from solo_server.utils.server_utils import start_vllm_server, setup_ollama_server, setup_llama_cpp_server, is_huggingface_repo, pull_model_from_huggingface
-from solo_server.utils.llama_cpp_utils import start_llama_cpp_server
+from solo_server.utils.server_utils import (start_vllm_server, 
+                                            start_ollama_server, 
+                                            start_llama_cpp_server, 
+                                            is_huggingface_repo, 
+                                            pull_model_from_huggingface)
 
 class ServerType(str, Enum):
     OLLAMA = "ollama"
@@ -19,27 +20,50 @@ class ServerType(str, Enum):
     LLAMACPP = "llama.cpp"
 
 def serve(
-    server: str = typer.Option("ollama", "--server", "-s", help="Server type (ollama, vllm, llama.cpp)"),
-    model: str = typer.Option(None, "--model", "-m", help="Model name or path"),
-    port: int = typer.Option(None, "--port", "-p", help="Port to run the server on")
+    server: Optional[str] = typer.Option(None, "--server", "-s", help="Server type (ollama, vllm, llama.cpp)"),
+    model: Optional[str] = typer.Option(None, "--model", "-m", help="""Model name or path. Can be:
+    - HuggingFace repo ID (e.g., 'meta-llama/Llama-3.2-1B-Instruct')
+    - Ollama model Registry (e.g., 'llama3.2')
+    - Local path to a model file (e.g., '/path/to/model.gguf')
+    If not specified, the default model from configuration will be used."""),
+    port: Optional[int] = typer.Option(None, "--port", "-p", help="Port to run the server on")
 ):
-    """Start a model server with the specified model"""
+    """Start a model server with the specified model.
     
-    # Get hardware info and GPU configuration
-    cpu_model, cpu_cores, memory_gb, gpu_vendor, gpu_model, gpu_memory, compute_backend, os_name = detect_hardware()
+    If no server is specified, uses the server type from configuration.
+    To set up your configuration, run 'solo setup' first.
+    """
     
-    # Load GPU configuration from config file
-    use_gpu = False
-    if os.path.exists(CONFIG_PATH):
-        with open(CONFIG_PATH, 'r') as f:
-            config = json.load(f)
-            use_gpu = config.get('hardware', {}).get('use_gpu', False)
+    # Check if config file exists
+    if not os.path.exists(CONFIG_PATH):
+        typer.echo("❌ Configuration file not found. Please run 'solo setup' first.", err=True)
+        typer.echo("Run 'solo setup' to complete the Solo Server setup and then try again.")
+        raise typer.Exit(code=1)
+    
+    # Load configuration
+    with open(CONFIG_PATH, 'r') as f:
+        config = json.load(f)
+    
+    # Extract hardware info from config
+    hardware_config = config.get('hardware', {})
+    use_gpu = hardware_config.get('use_gpu', False)
+    cpu_model = hardware_config.get('cpu_model')
+    gpu_vendor = hardware_config.get('gpu_vendor')
+    os_name = hardware_config.get('os')
+    
+    # If hardware info isn't in config, detect it
+    if not cpu_model or not gpu_vendor or not os_name:
+        cpu_model, _, _, gpu_vendor, _, _, _, os_name = detect_hardware()
     
     # Only enable GPU if configured and supported
     gpu_enabled = use_gpu and gpu_vendor in ["NVIDIA", "AMD", "Apple Silicon"]
     
-    # Normalize server name
-    server = server.lower()
+    # Use server from config if not specified
+    if not server:
+        server = config.get('server', {}).get('type', ServerType.OLLAMA.value)
+    else:
+        # Normalize server name
+        server = server.lower()
     
     # Validate server type
     if server not in [s.value for s in ServerType]:
@@ -62,29 +86,32 @@ def serve(
     
     if not port:
         if server == ServerType.VLLM.value:
-            port = vllm_config.get('default_port', 8000)
+            port = vllm_config.get('default_port', 5070)
         elif server == ServerType.OLLAMA.value:
-            port = ollama_config.get('default_port', 11434)
+            port = ollama_config.get('default_port', 5070)
         elif server == ServerType.LLAMACPP.value:
-            port = llama_cpp_config.get('default_port', 8080)
+            port = llama_cpp_config.get('default_port', 5070)
     
     # Start the appropriate server
+    typer.echo(f"\nStarting Solo server...")
+    success = False
+    original_model_name = model
+    server_pretty_name = server.capitalize()
+    
     if server == ServerType.VLLM.value:
         try:
-            if start_vllm_server(gpu_enabled, cpu_model, gpu_vendor, os_name, port, model):
-                typer.secho(
-                    f"✅ vLLM server is running at http://localhost:{port}\n" 
-                    f"Use 'docker logs -f {vllm_config.get('container_name', 'solo-vllm')}' to view the logs.",
-                    fg=typer.colors.BRIGHT_GREEN
-                )
+            success = start_vllm_server(gpu_enabled, cpu_model, gpu_vendor, os_name, port, model)
+            # Display container logs command
+            if success:
+                typer.echo(f"Use 'docker logs -f {vllm_config.get('container_name', 'solo-vllm')}' to view the logs.")
         except Exception as e:
-            typer.echo(f"❌ Failed to start vLLM server: {e}", err=True)
+            typer.echo(f"❌ Failed to start Solo Server: {e}", err=True)
             raise typer.Exit(code=1)
         
     elif server == ServerType.OLLAMA.value:
         # Start Ollama server
-        if not setup_ollama_server(gpu_enabled, gpu_vendor, port):
-            typer.echo("❌ Ollama server is not running!", err=True)
+        if not start_ollama_server(gpu_enabled, gpu_vendor, port):
+            typer.echo("❌ Failed to start Solo Server!", err=True)
             raise typer.Exit(code=1)
         
         # Pull the model if not already available
@@ -108,22 +135,27 @@ def serve(
                     ["docker", "exec", container_name, "ollama", "pull", model],
                     check=True
                 )
+            success = True
         except subprocess.CalledProcessError as e:
             typer.echo(f"❌ Failed to pull model: {e}", err=True)
             raise typer.Exit(code=1)
             
-        typer.secho(
-            f"✅ Ollama server is running at http://localhost:{port}",
-            fg=typer.colors.BRIGHT_GREEN
-        )
-        
     elif server == ServerType.LLAMACPP.value:
         # Start llama.cpp server with the specified model
-        if start_llama_cpp_server(os_name, model_path=model, port=port):
-            typer.secho(
-                f"✅ llama.cpp server is running at http://localhost:{port}",
-                fg=typer.colors.BRIGHT_GREEN
-            )
-        else:
-            typer.echo("❌ Failed to start llama.cpp server", err=True)
+        success = start_llama_cpp_server(os_name, model_path=model, port=port)
+        if not success:
+            typer.echo("❌ Failed to start Solo server", err=True)
             raise typer.Exit(code=1)
+    
+    # Display server information in the requested format
+    if success:
+        # Get formatted model name for display
+        display_model = original_model_name
+        if is_huggingface_repo(original_model_name):
+            # For HF models, get the repository name
+            display_model = original_model_name.split('/')[-1] if '/' in original_model_name else original_model_name
+        
+        # Print server information
+        typer.secho("✅ Solo Server is running", fg=typer.colors.BRIGHT_GREEN, bold=True)
+        typer.secho(f"Model  - {display_model}", fg=typer.colors.BRIGHT_BLUE, bold=True)
+        typer.secho(f"URL    - http://localhost:{port}", fg=typer.colors.BRIGHT_CYAN, bold=True)
