+++ solo_server/config/config.yaml
@@ -2,7 +2,7 @@
 servers:
   vllm:
     default_model: "meta-llama/Llama-3.2-1B-Instruct"
-    default_port: 8000
+    default_port: 5070
     max_model_len: 4096
     gpu_memory_utilization: 0.95
     container_name: "solo-vllm"
@@ -14,7 +14,7 @@ servers:
   
   ollama:
     default_model: "llama3.2"
-    default_port: 11434
+    default_port: 5070
     container_name: "solo-ollama"
     images:
       default: "ollama/ollama"
@@ -22,7 +22,7 @@ servers:
   
   llama_cpp:
     default_model: "bartowski/Llama-3.2-1B-Instruct-GGUF/llama-3.2-1B-Instruct-Q4_K_M.gguf"
-    default_port: 8080
+    default_port: 5070
     cmake_args:
       nvidia: "-DGGML_CUDA=on"
       amd: "-DGGML_HIPBLAS=on"
@@ -37,4 +37,4 @@ paths:
 # Timeouts
 timeouts:
   server_start: 30
-  docker_check: 20 
+  docker_check: 30
