+++ solo_server/utils/server_utils.py
@@ -4,11 +4,15 @@ import typer
 import click
 import sys
 import time
+import shutil
 import subprocess
 from solo_server.config import CONFIG_PATH
+from solo_server.utils.hf_utils import select_best_model_file
 from solo_server.config.config_loader import load_config, get_server_config, get_timeout_config
-from solo_server.utils.nvidia import is_cuda_toolkit_installed
-from solo_server.utils.llama_cpp_utils import is_uv_available, start_llama_cpp_server, get_available_models, select_best_model_file
+from solo_server.utils.llama_cpp_utils import (is_port_in_use, 
+                                              stop_server_on_port, 
+                                              preprocess_model_path, 
+                                              is_llama_cpp_installed)
 
 def start_vllm_server(gpu_enabled: bool, cpu: str = None, gpu_vendor: str = None, 
                       os_name:str = None, port: int = None, model: str = None):
@@ -23,7 +27,6 @@ def start_vllm_server(gpu_enabled: bool, cpu: str = None, gpu_vendor: str = None
     container_name = vllm_config.get('container_name', 'solo-vllm')
     
     # Initialize container_exists flag
-    typer.echo("Starting Solo server with vLLM...")
     container_exists = False
     try:
         # Check if container exists (running or stopped)
@@ -84,7 +87,7 @@ def start_vllm_server(gpu_enabled: bool, cpu: str = None, gpu_vendor: str = None
                     model_name = model
                     
                     # Get HuggingFace token from environment variable or config file
-                    typer.echo("\nChecking for HuggingFace token...")
+                    typer.echo("🔑 Checking for HuggingFace token...")
                     hf_token = os.getenv('HUGGING_FACE_TOKEN', '')
 
                     if not hf_token:  # If not in env, try config file
@@ -145,7 +148,7 @@ def start_vllm_server(gpu_enabled: bool, cpu: str = None, gpu_vendor: str = None
             ).stdout.strip()
 
             if not image_exists:
-                typer.echo(f"❌ vLLM server is not setup. Please run 'solo setup' first to setup vLLM.", err=True)
+                typer.echo(f"❌ Solo server is not setup. Please run 'solo setup' first.", err=True)
                 return False
 
             docker_run_cmd.append(image)
@@ -173,14 +176,13 @@ def start_vllm_server(gpu_enabled: bool, cpu: str = None, gpu_vendor: str = None
 
             if gpu_vendor == "NVIDIA":
                 # Get GPU memory utilization from config
-                gpu_memory_utilization = vllm_config.get('gpu_memory_utilization', 0.95)
+                gpu_memory_utilization = vllm_config.get('gpu_memory_utilization', 0.85)
                 docker_run_cmd += [
                     "--gpu_memory_utilization", str(gpu_memory_utilization)
                 ]
                 if 5 < compute_cap < 8:
                     docker_run_cmd += ["--dtype", "half"]
-        
-            typer.echo("Starting Solo server with vLLM...")
+
             subprocess.run(docker_run_cmd, check=True, capture_output=True)
             
             # Check docker logs for any errors
@@ -225,7 +227,7 @@ def start_vllm_server(gpu_enabled: bool, cpu: str = None, gpu_vendor: str = None
         typer.echo(f"❌ Unexpected error: {e}", err=True)
         return False
 
-def setup_ollama_server(gpu_enabled: bool = False, gpu_vendor: str = None, port: int = None):
+def start_ollama_server(gpu_enabled: bool = False, gpu_vendor: str = None, port: int = None):
     """Setup solo-server Ollama environment."""
     # Load Ollama configuration from YAML
     ollama_config = get_server_config('ollama')
@@ -246,7 +248,6 @@ def setup_ollama_server(gpu_enabled: bool = False, gpu_vendor: str = None, port:
             text=True
         ).stdout.strip()
 
-        typer.echo(f"Starting Solo server with Ollama...")
         if container_exists:
             # Check if container is running
             check_cmd = ["docker", "ps", "-q", "-f", f"name={container_name}"]
@@ -281,11 +282,11 @@ def setup_ollama_server(gpu_enabled: bool = False, gpu_vendor: str = None, port:
             try:
                 subprocess.run(["docker", "image", "inspect", image], check=True, capture_output=True)
             except subprocess.CalledProcessError:
-                typer.echo("❌ Ollama is not setup. Please run 'solo setup' first", err=True)
+                typer.echo("❌ Solo server is not setup. Please run 'solo setup' first", err=True)
                 return False
 
             # Start Ollama container
-            docker_run_cmd = ["docker", "run", "-d", "--name", container_name, "-v", "ollama:/root/.ollama", "-p", f"{port}:{port}"]
+            docker_run_cmd = ["docker", "run", "-d", "--name", container_name, "-v", "ollama:/root/.ollama", "-p", f"{port}:11434"]
             if gpu_vendor == "NVIDIA" and gpu_enabled:
                 docker_run_cmd += ["--gpus", "all"]
             elif gpu_vendor == "AMD" and gpu_enabled:
@@ -319,6 +320,125 @@ def setup_ollama_server(gpu_enabled: bool = False, gpu_vendor: str = None, port:
         raise typer.Exit(code=1)
     except Exception as e:
         typer.echo(f"❌ Unexpected error: {e}", err=True)
+        return False
+    
+
+def start_llama_cpp_server(os_name: str = None, model_path: str = None, port: int = None):
+    """
+    Start the llama.cpp server.
+    
+    Parameters:
+    os_name (str, optional): The name of the operating system.
+    model_path (str, optional): Path to the model file or HuggingFace repo ID.
+    port (int, optional): Port to run the server on.
+    """
+    # Check if llama_cpp is installed
+    if not is_llama_cpp_installed():
+        typer.echo("❌ Server not found. Please run 'solo setup' first.", err=True)
+        return False
+        
+    # Load llama.cpp configuration from YAML
+    llama_cpp_config = get_server_config('llama_cpp')
+    
+    # Use default values from config if not provided
+    port = port or llama_cpp_config.get('default_port', 8080)
+    model_path = model_path or llama_cpp_config.get('default_model')
+    
+    try:
+        # Check if port is already in use
+        if is_port_in_use(port):
+            typer.echo(f"Port {port} is already in use.")
+            if not stop_server_on_port(port):
+                typer.echo(f"❌ Failed to stop existing server on port {port}. Please stop it manually.", err=True)
+                return False
+        
+        # If no model path is provided, prompt the user
+        if not model_path:
+            typer.echo("Please provide the path to your GGUF model file or a HuggingFace repo ID.")
+            model_path = typer.prompt("Enter the model path or repo ID")
+            
+        # Get HuggingFace token if needed
+        hf_token = os.getenv('HUGGING_FACE_TOKEN', '')
+        if not hf_token and not os.path.exists(model_path):  # Only check for token if not a local file
+            if os.path.exists(CONFIG_PATH):
+                with open(CONFIG_PATH, 'r') as f:
+                    config = json.load(f)
+                    hf_token = config.get('hugging_face', {}).get('token', '')
+        
+        # Determine if this is a repo ID or direct path
+        hf_repo_id, model_pattern = preprocess_model_path(model_path, hf_token)
+
+        # Build server command
+        server_cmd = [
+            sys.executable, "-m", "llama_cpp.server",
+            "--host", "0.0.0.0",
+            "--port", str(port)
+        ]
+        
+        if hf_repo_id and not os.path.exists(model_path):
+            # This is a HuggingFace repo ID
+            typer.echo(f"Using HuggingFace repo: {hf_repo_id}")
+            server_cmd.extend(["--hf_model_repo_id", hf_repo_id])
+            server_cmd.extend(["--model", model_pattern])
+        else:
+            # Direct model path
+            model_path = os.path.abspath(os.path.expanduser(model_path))
+            if not os.path.exists(model_path):
+                typer.echo(f"❌ Model file not found: {model_path}", err=True)
+                return False
+            server_cmd.extend(["--model", model_path])
+        
+        if os_name == "Windows":
+            # Create a log file for capturing output
+            log_dir = os.path.join(os.path.expanduser("~"), ".solo", "logs")
+            os.makedirs(log_dir, exist_ok=True)
+            log_file = os.path.join(log_dir, "llama_cpp_server.log")
+            
+            # Start the server in a new console window and keep it open with a pause command
+            cmd_str = " ".join(server_cmd) + " & pause"
+            process = subprocess.Popen(
+                f'start cmd /k "{cmd_str}"',
+                shell=True
+            )
+            typer.echo(f"Solo Server is running in a new terminal window. Use ctrl + c to stop.")
+        else:
+            # For Unix-like systems, use terminal-specific commands
+            if os_name == "Darwin":  # macOS
+                # For macOS, use AppleScript to keep the Terminal window open
+                script = f'tell app "Terminal" to do script "{" ".join(server_cmd)} ; echo \'\\nServer is running. Press Ctrl+C to stop.\'; bash"'
+                terminal_cmd = ["osascript", "-e", script]
+                subprocess.Popen(terminal_cmd)
+                typer.echo("Server is running in a new Terminal window")
+            else:  # Linux and other Unix-like systems
+                # Try to detect the terminal and keep it open
+                if shutil.which("gnome-terminal"):
+                    terminal_cmd = ["gnome-terminal", "--", "bash", "-c", f"{' '.join(server_cmd)}; echo '\\nServer is running. Press Ctrl+C to stop.'; exec bash"]
+                    subprocess.Popen(terminal_cmd)
+                elif shutil.which("xterm"):
+                    terminal_cmd = ["xterm", "-e", f"{' '.join(server_cmd)}; echo '\\nServer is running. Press Ctrl+C to stop.'; exec bash"]
+                    subprocess.Popen(terminal_cmd)
+                elif shutil.which("konsole"):
+                    terminal_cmd = ["konsole", "-e", f"bash -c '{' '.join(server_cmd)}; echo \"\\nServer is running. Press Ctrl+C to stop.\"; exec bash'"]
+                    subprocess.Popen(terminal_cmd)
+                else:
+                    # Fallback to background process if no terminal is found
+                    process = subprocess.Popen(
+                        server_cmd,
+                        stdout=subprocess.PIPE,
+                        stderr=subprocess.PIPE,
+                        start_new_session=True
+                    )
+                    typer.echo(f"Server is running in the background. Process ID: {process.pid}")
+                    return True
+                
+                typer.echo("Server is running in a new terminal window")
+        
+        # Wait for the server to start
+        time.sleep(2)
+        return True
+        
+    except Exception as e:
+        typer.echo(f"❌ Failed to start Solo server: {e}", err=True)
         return False
 
 def is_huggingface_repo(model: str) -> bool:
@@ -424,94 +544,3 @@ def pull_model_from_huggingface(container_name: str, model: str) -> str:
     except subprocess.CalledProcessError as e:
         typer.echo(f"❌ Failed to pull model from HuggingFace: {e}", err=True)
         raise e
-
-def setup_llama_cpp_server(gpu_enabled: bool, gpu_vendor: str = None, os_name: str = None, install_only: bool = False):
-    """
-    Setup llama_cpp_python server using system config.
-
-    Parameters:
-    gpu_enabled (bool): Whether GPU is enabled.
-    gpu_vendor (str, optional): The GPU vendor (e.g., NVIDIA, AMD, Apple Silicon).
-    os_name (str, optional): The name of the operating system.
-    install_only (bool, optional): If True, only install the library without starting the server.
-    """
-    # Load llama.cpp configuration from YAML
-    llama_cpp_config = get_server_config('llama_cpp')
-    
-    # Check if llama-cpp-python is already installed
-    try:
-        import llama_cpp
-        typer.echo("Starting Solo server with llama.cpp...")
-        if install_only:
-            return True
-        return start_llama_cpp_server(os_name)
-    except ImportError:
-        typer.echo("Installing llama.cpp server...")
-
-    # Set CMAKE_ARGS based on hardware and OS
-    cmake_args = []
-    use_gpu_build = False
-    
-    if gpu_enabled:
-        if gpu_vendor == "NVIDIA":
-            if is_cuda_toolkit_installed():
-                use_gpu_build = True
-                cmake_args.append(llama_cpp_config.get('cmake_args', {}).get('nvidia', "-DGGML_CUDA=on"))
-            else:
-                typer.echo("⚠️ NVIDIA CUDA Toolkit not properly configured. Will try CPU-only build instead.", err=True)
-        elif gpu_vendor == "AMD":
-            use_gpu_build = True
-            cmake_args.append(llama_cpp_config.get('cmake_args', {}).get('amd', "-DGGML_HIPBLAS=on"))
-        elif gpu_vendor == "Apple Silicon":
-            use_gpu_build = True
-            cmake_args.append(llama_cpp_config.get('cmake_args', {}).get('apple_silicon', "-DGGML_METAL=on"))
-  
-    cmake_args_str = " ".join(cmake_args)
-
-    try:
-        env = os.environ.copy()
-        if use_gpu_build:
-            env["CMAKE_ARGS"] = cmake_args_str
-            typer.echo(f"Attempting GPU-accelerated build with: {cmake_args_str}")
-        else:
-            typer.echo("Installing CPU-only version of llama-cpp-python")
-        
-        # Install llama-cpp-python using the Python interpreter
-        if is_uv_available():
-            use_uv = typer.confirm("uv is available. Are you using (uv's) virtual env for installation?", default=False)
-            if use_uv:
-                installer_cmd = ["uv", "pip", "install", "--no-cache-dir", "llama-cpp-python[server]"]
-            else:
-                installer_cmd = [sys.executable, "-m", "pip", "install", "--no-cache-dir", "llama-cpp-python[server]"]
-        else:
-            installer_cmd = [sys.executable, "-m", "pip", "install", "--no-cache-dir", "llama-cpp-python[server]"]
-
-        try:
-            subprocess.check_call(installer_cmd, env=env)
-        except subprocess.CalledProcessError as e:
-            if use_gpu_build:
-                typer.echo("❌ GPU-accelerated build failed. Falling back to CPU-only build...", err=True)
-                # Clear CMAKE_ARGS for CPU-only build
-                env.pop("CMAKE_ARGS", None)
-                subprocess.check_call(installer_cmd, env=env)
-            else:
-                raise e
-        
-        if install_only:
-            return True
-            
-        # Start the server if installation was successful
-        try:
-            if start_llama_cpp_server(os_name):
-                typer.echo("\n ✅ llama.cpp server is ready!")
-                return True
-            else:
-                return False
-        except Exception as e:
-            typer.echo(f"❌ Failed to start llama.cpp server: {e}", err=True)
-            return False
-
-    except subprocess.CalledProcessError as e:
-        typer.echo(f"❌ Failed to setup llama.cpp server: {e}", err=True)
-        return False
-
